Add snowflake for a person functionality
Added the ability for user to input first and last names plus date of
birth to generate a unique snowflake for that person.

That necessitated implementation of a seedable PRNG and string to hash
function. Luckily the internet provided nice public domain
implementations of both.

See: https://stackoverflow.com/questions/521295